Change string representation of empty Optional
An empty Optional currently has the same string representation as a non-empty Optional with a zero value. This is detrimental in terms of representation as it's impossible to distinguish the two cases and will likely lead to confusion.

This commit resolves this by following Go's pattern for `<nil>` being used to represent nil values by having `Optional.String` return `<empty>` when the Optional has no value.